Convert bluepill_service to a custom resource
Stop storing state as attributes
Fix converge_by messaging if the user is specifying a custom service name
Remove the supports attribute which wasn’t being used (for chef 13 compatibility)
Remove references to using a service resource and specifying the provider. Please don’t do this.
