+++ .kitchen.yml
@@ -9,15 +9,11 @@ platforms:
   - name: centos-6.8
   - name: centos-7.3
   - name: debian-7.11
-    run_list: apt::default
   - name: debian-8.6
-    run_list: apt::default
   - name: fedora-25
-  - name: opensuse-leap-42.2
+  - name: opensuse-leap-42.1
   - name: ubuntu-14.04
-    run_list: apt::default
   - name: ubuntu-16.04
-    run_list: apt::default
   - name: freebsd-10.3
   - name: freebsd-11.0
