+++ Berksfile
@@ -3,6 +3,5 @@ source 'https://supermarket.chef.io'
 metadata
 
 group :integration do
-  cookbook 'apt'
   cookbook 'bluepill_test', path: './test/cookbooks/bluepill_test'
 end
