+++ README.md
@@ -1,20 +1,27 @@
 # bluepill Cookbook
+
 [![Build Status](https://travis-ci.org/chef-cookbooks/bluepill.svg?branch=master)](https://travis-ci.org/chef-cookbooks/bluepill) [![Cookbook Version](https://img.shields.io/cookbook/v/bluepill.svg)](https://supermarket.chef.io/cookbooks/bluepill)
 
 Installs bluepill Ruby Gem and configures it to manage services. Also includes a LWRP.
 
 ## Requirements
+
 ### Platforms
+
 Bluepill is a pure Ruby service management tool/library, so this cookbook should work on any system. The attributes do set up paths based on FHS locations, see below.
 
 ### Chef
-- Chef 12.1+
+
+- Chef 12.5+
 
 ### Cookbooks
+
 - none
 
 ## Attributes
+
 Default locations for bluepill are in "FHS compliant" locations.
+
 - `node["bluepill"]["bin"]` - Path to bluepill program, default is 'bluepill' in the RubyGems binary directory.
 - `node["bluepill"]["logfile"]` - Location of the bluepill log file, default "/var/log/bluepill.log".
 - `node["bluepill"]["conf_dir"]` - Location of service config files (pills), default "/etc/bluepill".
@@ -24,30 +31,25 @@ Default locations for bluepill are in "FHS compliant" locations.
 - `node["bluepill"]["version"]` - Version of bluepill to install, default is latest.
 - `node["bluepill"]["use_rsyslog"]` - Enable configuration and use of rsyslog for bluepill.
 
-# Custom Resources
-This cookbook contains an LWRP, `bluepill_service`. This can be used with the normal Chef service resource, by using the `provider` parameter, or by specifying the `bluepill_service` shortcut. These two resources are equivalent.
+# Resources
+
+This cookbook contains a resource, `bluepill_service`. This can be used with the normal Chef service resource, by using the `provider` parameter, or by specifying the `bluepill_service` shortcut. These two resources are equivalent.
 
 ```ruby
-service 'my_app' do
-  provider bluepill_service
-  action [:enable, :load, :start]
-end
-
 bluepill_service 'my_app' do
   action [:enable, :load, :start]
 end
 ```
 
-The load action should probably always be specified, to ensure that if bluepill isn't running already it gets started. The
-
-The recipe using the service must contain a template resource for the pill and it must be named `my_app.pill.erb`, where `my_app` is the service name passed to the bluepill service resource.
+The load action should probably always be specified, to ensure that if bluepill isn't running already it gets started. The The recipe using the service must contain a template resource for the pill and it must be named `my_app.pill.erb`, where `my_app` is the service name passed to the bluepill service resource.
 
 ## Usage
+
 Be sure to include the bluepill recipe in the run list to ensure that the gem and bluepill-related directories are created. This will also make the cookbook available on the system and other cookbooks won't need to explicitly depend on it in the metadata.
 
 If the default directory locations in the attributes/default.rb aren't what you want, change them by setting them either in the attributes file itself, or create attributes in a role applied to any systems that will use bluepill.
 
-Bluepill uses configuration files named my_app.pill that are stored in `/etc/bluepill`.  Here is an example pill:
+Bluepill uses configuration files named my_app.pill that are stored in `/etc/bluepill`. Here is an example pill:
 
 ```ruby
 Bluepill.application('my_app') do |app|
@@ -57,6 +59,7 @@ Bluepill.application('my_app') do |app|
   end
 end
 ```
+
 You can create this file with `cookbook_file` as follows:
 
 ```ruby
@@ -68,6 +71,7 @@ end
 See bluepill's [documentation](https://github.com/bluepill-rb/bluepill#config) for more information on creating pill templates.
 
 ## Testing
+
 This cookbook has the following [ChefSpec custom matchers](https://github.com/sethvargo/chefspec#packaging-custom-matchers) defined:
 
 - enable_bluepill_service
@@ -95,6 +99,7 @@ end
 ```
 
 ## License & Authors
+
 **Author:** Cookbook Engineering Team ([cookbooks@chef.io](mailto:cookbooks@chef.io))
 
 **Copyright:** 2010-2016, Chef Software, Inc.
@@ -112,4 +117,3 @@ WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 See the License for the specific language governing permissions and
 limitations under the License.
 ```
-
