+++ attributes/default.rb
@@ -1,7 +1,7 @@
 # Cookbook:: bluepill
 # Attributes:: default
 #
-# Copyright:: 2010-2016, Chef Software, Inc.
+# Copyright:: 2010-2017, Chef Software, Inc.
 #
 # Licensed under the Apache License, Version 2.0 (the "License");
 # you may not use this file except in compliance with the License.
