+++ recipes/rsyslog.rb
@@ -2,7 +2,7 @@
 # Cookbook:: bluepill
 # Recipe:: rsyslog
 #
-# Copyright:: 2010-2016, Chef Software, Inc.
+# Copyright:: 2010-2017, Chef Software, Inc.
 #
 # Licensed under the Apache License, Version 2.0 (the "License");
 # you may not use this file except in compliance with the License.
