+++ test/cookbooks/bluepill_test/metadata.rb
@@ -1,5 +1,4 @@
 name 'bluepill_test'
 license 'Apache 2.0'
 version '0.0.1'
-
 depends 'bluepill'
